backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.0.0

pytest==7.1.2
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3

flake8==6.0.0
flake8-isort==6.0.0

gunicorn==20.1.0
